Add methods to compute and save models, fields and records to database
Add few methods to be allowed to compute and save models, fields and/or records to the database

+++ code_gen/src/model/field.rs
@@ -64,7 +64,7 @@ impl FieldGen {
                             }
                         },
                         Lit::Float(f) => {
-                            let float = f.base10_parse::<f64>();
+                            let float = f.base10_parse::<f32>();
                             if float.is_ok() {
                                 FieldType::Float(float?)
                             } else {
